requirements.txt
Django==3.1.2
requests==2.25.0
beautifulsoup4==4.9.3
lxml
#gunicorn
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-crontab==0.7.1